Changes made to test system

+++ MoneyEntry.DataAccess.EFCore/Expenses/Models/Views/vTrans.cs
@@ -9,7 +9,7 @@ namespace MoneyEntry.DataAccess.EFCore.Expenses.Models
     {
         public int PersonID { get; set; }
         public string Name { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string TransactionDesc { get; set; }
         public byte TypeID { get; set; }
         public string Type { get; set; }
@@ -18,7 +18,7 @@ namespace MoneyEntry.DataAccess.EFCore.Expenses.Models
         public DateTime CreatedDate { get; set; }
         [Key]
         public int TransactionID { get; set; }
-        public double RunningTotal { get; set; }
+        public decimal RunningTotal { get; set; }
         public bool reconciled { get; set; }
     }
 }
